Add GoalMailer and EmailGoalAffirmersService
Emails are sent to affirmers and owner of a goal that has met its
critical mass. This is achieved through `GoalMailer` and
`EmailGoalAffirmersService`. After a user affirms a goal, we check to
see if the affirmation count is equal to the critical mass (this
'prevents' us from sending email when the goal is over critical mass).