Add omniauth_facebook and relevant migrations
When a new user tries to sign in with facebook, a new account is
created. When an old user tries to log in with facebook but does not
have a connected facebook account, he is logged in, and facebook
credentials are stored. If the user has an account and has already
connected, he is logged in.